.github/actions/deploy-s3-docker/requirements.txt
PyGithub>=1.55
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12